New custom three.js, bugfix
- upped versions to node 3->4, npm 3->5 and upped all other js/npm dependencies
- Added build 'three' in gulpfile which will build a custom minimized three.js lib with only the components defined in javascript/lib/ threeSmall.js
- New three.js file is smaller and there is now proper js autocomplete for three.js in VSCode thanks to d.ts  file
- The lib is compiled to resources/compiled and used like other non-modular libs in resources/compiled/lib
- refactored and fixed AssetService#onAssetTaskCreation

+++ src/main/java/gmm/service/assets/AssetService.java
@@ -155,26 +155,26 @@ public class AssetService {
 		
 		final AssetTaskService<A> service = serviceFinder.getAssetService(Util.classOf(task));
 		
-		for (final AssetGroupType type : AssetGroupType.values()) {
-			final AssetInfo info = getAssetInfo(name, type);
+		{
+			final AssetGroupType type = AssetGroupType.ORIGINAL;
+			final OriginalAssetFileInfo info = getOriginalAssetFileInfo(name);
 			final AssetProperties props = task.getAssetProperties(type);
-			if (info != null) {
-				boolean hasAsset = true;
-				if (info instanceof NewAssetFolderInfo) {
-					final NewAssetFolderInfo folderInfo = (NewAssetFolderInfo) info;
-					if (folderInfo.getStatus() != AssetFolderStatus.VALID_WITH_ASSET) {
-						service.removeAssetProperties(task, type);
-						hasAsset = false;
-					}
-				}
-				if (hasAsset && (props == null || !service.isValidAssetProperties(props, info))) {
-					service.recreateAssetProperties(task, info);
-				}
-			} else {
-				if (props != null) {
-					service.removeAssetProperties(task, type);
-				}
+			
+			if (info != null && props == null) service.recreateAssetProperties(task, info);
+			if (info == null && props != null) service.removeAssetProperties(task, type);
+		}{
+			final AssetGroupType type = AssetGroupType.NEW;
+			final NewAssetFolderInfo info = getNewAssetFolderInfo(name);
+			final AssetProperties props = task.getAssetProperties(type);
+			
+			final boolean existsAndValid =
+					(info != null) && (info.getStatus() == AssetFolderStatus.VALID_WITH_ASSET);
+			
+			if (existsAndValid && props == null) service.recreateAssetProperties(task, info);
+			if (existsAndValid && props != null && !service.isValidAssetProperties(props, info)) {
+				service.recreateAssetProperties(task, info);
 			}
+			if (!existsAndValid && props != null) service.removeAssetProperties(task, type);
 		}
 	}
